remove arethusa from non-embedded treebanks
