+++ package.json
@@ -26,7 +26,7 @@
     "react-scripts": "3.4.1",
     "react-test-renderer": "^16.13.1",
     "reactstrap": "^8.5.1",
-    "treebank-react": "^1.0.1"
+    "treebank-react": "^1.1.0"
   },
   "scripts": {
     "predeploy": "yarn run build",
