+++ src/components/ControlPanel/ControlPanel.js
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import { SettingsIcon } from '@primer/octicons-react';
-import queryString from 'query-string';
 
-import { chunksType, publicationMatchType, queryType } from '../../lib/types';
+import { chunksType, publicationMatchType } from '../../lib/types';
 
 import styles from './ControlPanel.module.css';
 
@@ -22,23 +20,55 @@ const getFbcnlFromNumbers = (chunk, numbers) => {
   ];
 };
 
+const renderBack = (first, back, current) => {
+  const visibility = current === String(first) ? ' invisible' : '';
+
+  return (
+    <>
+      <li className={`nav-item${visibility}`}>
+        <Link className={`nav-link text-light ${styles.link}`} to={`./${first}`}>
+          &laquo; First
+        </Link>
+      </li>
+      <li className={`nav-item${visibility}`}>
+        <Link className={`nav-link text-light ${styles.link}`} to={`./${back}`}>
+          &#8249; Back
+        </Link>
+      </li>
+    </>
+  );
+};
+
+const renderNext = (current, next, last) => {
+  const visibility = current === String(last) ? ' invisible' : '';
+
+  return (
+    <>
+      <li className={`nav-item${visibility}`}>
+        <Link className={`nav-link text-light ${styles.link}`} to={`./${next}`}>
+          Next &#8250;
+        </Link>
+      </li>
+      <li className={`nav-item${visibility}`}>
+        <Link className={`nav-link text-light ${styles.link}`} to={`./${last}`}>
+          Last &raquo;
+        </Link>
+      </li>
+    </>
+  );
+};
+
 class ControlPanel extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
       refIsOpen: false,
-      settingsIsOpen: false,
     };
 
     this.getLines = this.getLines.bind(this);
     this.getFbcnl = this.getFbcnl.bind(this);
-    this.createLink = this.createLink.bind(this);
     this.toggleRefOpen = this.toggleRefOpen.bind(this);
-    this.toggleSettingsOpen = this.toggleSettingsOpen.bind(this);
-    this.renderSettingsLinks = this.renderSettingsLinks.bind(this);
-    this.renderBack = this.renderBack.bind(this);
-    this.renderNext = this.renderNext.bind(this);
   }
 
   getLines() {
@@ -74,92 +104,20 @@ class ControlPanel extends Component {
     ];
   }
 
-  createLink(to) {
-    const { linkQuery } = this.props;
-    const link = `./${to}`;
-
-    if (Object.entries(linkQuery).length > 0) {
-      const query = queryString.stringify(linkQuery);
-
-      return `${link}?${query}`;
-    }
-
-    return link;
-  }
-
   toggleRefOpen() {
     this.setState(({ refIsOpen }) => ({ refIsOpen: !refIsOpen }));
   }
 
-  toggleSettingsOpen() {
-    this.setState(({ settingsIsOpen }) => ({ settingsIsOpen: !settingsIsOpen }));
-  }
-
-  renderSettingsLinks() {
-    const { fullQuery } = this.props;
-    const { config } = fullQuery;
-    const newConfig = config === 'sidepanel' ? 'default' : 'sidepanel';
-    const text = config === 'sidepanel' ? 'Hide morphology' : 'Show morphology';
-
-    return (
-      <a
-        href={`?${queryString.stringify({ ...fullQuery, config: newConfig })}`}
-        className="dropdown-item"
-      >
-        {text}
-      </a>
-    );
-  }
-
-  renderBack(first, back, current) {
-    const visibility = current === String(first) ? ' invisible' : '';
-
-    return (
-      <>
-        <li className={`nav-item${visibility}`}>
-          <Link className={`nav-link text-light ${styles.link}`} to={this.createLink(first)}>
-            &laquo; First
-          </Link>
-        </li>
-        <li className={`nav-item${visibility}`}>
-          <Link className={`nav-link text-light ${styles.link}`} to={this.createLink(back)}>
-            &#8249; Back
-          </Link>
-        </li>
-      </>
-    );
-  }
-
-  renderNext(current, next, last) {
-    const visibility = current === String(last) ? ' invisible' : '';
-
-    return (
-      <>
-        <li className={`nav-item${visibility}`}>
-          <Link className={`nav-link text-light ${styles.link}`} to={this.createLink(next)}>
-            Next &#8250;
-          </Link>
-        </li>
-        <li className={`nav-item${visibility}`}>
-          <Link className={`nav-link text-light ${styles.link}`} to={this.createLink(last)}>
-            Last &raquo;
-          </Link>
-        </li>
-      </>
-    );
-  }
-
   render() {
-    const { refIsOpen, settingsIsOpen } = this.state;
+    const { refIsOpen } = this.state;
     const [first, back, current, next, last] = this.getFbcnl();
     const lines = this.getLines();
 
     return (
       <nav className="navbar navbar-expand navbar-dark bg-dark">
         <div className="collapse navbar-collapse" id="controlPanel">
-          <ul className={`navbar-nav mr-auto ${styles.dummyIcon}`} />
           <ul className="navbar-nav mx-auto">
-            {this.renderBack(first, back, current)}
+            {renderBack(first, back, current)}
             <li className="nav-item dropdown">
               <button className={`btn btn-link nav-link text-light dropdown-toggle ${styles.dropdownButton}`} type="button" aria-haspopup="true" aria-expanded={refIsOpen} onClick={this.toggleRefOpen}>
                 {current}
@@ -167,24 +125,14 @@ class ControlPanel extends Component {
               <div className={`dropdown-menu ${styles.dropdownScroll} ${refIsOpen ? 'show' : ''}`}>
                 {
                   lines.map((n) => (
-                    <Link className="dropdown-item" key={n} to={this.createLink(n)} onClick={this.toggleRefOpen}>
+                    <Link className="dropdown-item" key={n} to={`./${n}`} onClick={this.toggleRefOpen}>
                       {n}
                     </Link>
                   ))
                 }
               </div>
             </li>
-            {this.renderNext(current, next, last)}
-          </ul>
-          <ul className="navbar-nav ml-auto">
-            <li className="nav-item dropdown dropleft">
-              <button className="btn btn-link nav-link text-light" type="button" aria-haspopup="true" aria-expanded={settingsIsOpen} onClick={this.toggleSettingsOpen}>
-                <SettingsIcon />
-              </button>
-              <div className={`dropdown-menu ${styles.dropdownScroll} ${settingsIsOpen ? 'show' : ''}`}>
-                {this.renderSettingsLinks()}
-              </div>
-            </li>
+            {renderNext(current, next, last)}
           </ul>
         </div>
       </nav>
@@ -195,8 +143,6 @@ class ControlPanel extends Component {
 ControlPanel.propTypes = {
   chunks: chunksType.isRequired,
   match: publicationMatchType.isRequired,
-  fullQuery: queryType.isRequired,
-  linkQuery: queryType.isRequired,
 };
 
 export default ControlPanel;
