+++ src/components/PublicationDirector/PublicationDirector.js
@@ -10,7 +10,7 @@ class PublicationDirector extends Component {
 
     const { config } = props;
     const argsLookup = {};
-    const { logo, link, treebankReact } = config;
+    const { logo, link } = config;
 
     config.collections.forEach((collection) => {
       (collection.publications || []).forEach((publication) => {
@@ -32,7 +32,6 @@ class PublicationDirector extends Component {
           argsLookup[path] = {
             logo,
             link,
-            treebankReact,
             publicationPath,
             author,
             work,
