+++ src/components/Treebank/Treebank.module.css
@@ -1,7 +1,3 @@
-.treebankContainer {
-  position: relative;
-}
-
 .text p {
   margin-bottom: 0.7rem;
   margin-top: 0.7rem;
