+++ src/config.json
@@ -6,7 +6,6 @@
   "report": "https://github.com/perseids-publications/treebank-template/issues",
   "github": "https://github.com/perseids-publications/treebank-template/",
   "twitter": "https://twitter.com/PerseidsProject",
-  "treebankReact": true,
   "collections": [
     {
       "title": "Information",
