+++ src/lib/params.js
@@ -2,7 +2,6 @@ import queryString from 'query-string';
 
 const params = [
   { name: 'w', type: 'array' },
-  { name: 'config', type: 'string', link: true },
 ];
 
 const convertToType = (value, type) => {
@@ -20,19 +19,6 @@ export const parse = (search) => {
   params.forEach(({ name, type }) => {
     if (Object.prototype.hasOwnProperty.call(parsed, name)) {
       result[name] = convertToType(parsed[name], type);
-    }
-  });
-
-  return result;
-};
-
-export const linkParams = (search) => {
-  const additionalArgs = parse(search);
-  const result = {};
-
-  params.forEach(({ name, type, link }) => {
-    if (link && Object.prototype.hasOwnProperty.call(additionalArgs, name)) {
-      result[name] = convertToType(additionalArgs[name], type);
     }
   });
